Refactor 404 not found pages and routes
